creates component to automatically switch from preview to content when launch date is hit #201
