+++ src/components/project-pages/EroticStoriesPreview.tsx
@@ -21,13 +21,15 @@ const Img = styled.img<LayoutProps>`
   ${layout};
 `;
 
-const EroticStoriesPreview = () => {
+const EroticStoriesPreview: React.FC<{ launchDate: string }> = ({
+  launchDate,
+}) => {
   return (
     <Main justifyContent="center" alignItems="center">
       <Container justifyContent="center" alignItems="center">
         <Img src={statue} alt="statue icon" maxWidth="30%" />
       </Container>
-      <Timer endDate="2020-12-21" />
+      <Timer launchDate={launchDate} />
     </Main>
   );
 };
