+++ src/components/project-pages/Eye.tsx
@@ -1,8 +1,9 @@
-import React from "react";
+import React, { Fragment } from "react";
 import styled from "styled-components";
 import Timer from "../Timer";
 import { FlexboxProps, flexbox, layout, LayoutProps } from "styled-system";
 import eye from "../assets/project-page/eye.png";
+import PreviewOrProjectPage from "./PreviewOrProjectPage";
 
 const Main = styled.main<FlexboxProps>`
   display: flex;
@@ -20,14 +21,25 @@ const Container = styled.div<FlexboxProps>`
 const Img = styled.img<LayoutProps>`
   ${layout};
 `;
-
-const Eye = () => {
+const PreviewPage: React.FC<{ launchDate: string }> = ({ launchDate }) => {
   return (
-    <Main justifyContent="center" alignItems="center">
+    <Fragment>
       <Container justifyContent="center" alignItems="center">
         <Img src={eye} alt="eye icon" maxWidth="30%" />
       </Container>
-      <Timer endDate="2020-12-21" />
+      <Timer launchDate={launchDate} />
+    </Fragment>
+  );
+};
+const launchDate = "2019-12-21";
+const Eye = () => {
+  return (
+    <Main justifyContent="center" alignItems="center">
+      <PreviewOrProjectPage
+        launchDate={launchDate}
+        PreviewPage={PreviewPage}
+        ProjectPage={() => null}
+      />
     </Main>
   );
 };
