+++ src/components/project-pages/FashionEditorialPreview.tsx
@@ -4,7 +4,6 @@ import Timer from "../Timer";
 import { FlexboxProps, flexbox, layout, LayoutProps } from "styled-system";
 import stairs from "../assets/project-page/stairs.png";
 
-
 const Container = styled.div<FlexboxProps>`
   display: flex;
   ${flexbox};
@@ -14,13 +13,13 @@ const Img = styled.img<LayoutProps>`
   ${layout};
 `;
 
-const FashionEditorial = () => {
+const FashionEditorial: React.FC<{ launchDate: string }> = ({ launchDate }) => {
   return (
     <Fragment>
       <Container justifyContent="center" alignItems="center">
         <Img src={stairs} alt="stairs icon" maxWidth="30%" />
       </Container>
-      <Timer endDate="2020-11-20" />
+      <Timer launchDate={launchDate} />
     </Fragment>
   );
 };
