+++ src/components/project-pages/KaiLandre.tsx
@@ -1,8 +1,10 @@
 import React from "react";
 import styled from "styled-components";
 import { FlexboxProps, flexbox, layout, LayoutProps } from "styled-system";
+import ReactPlayer from "react-player/lazy";
+
 import KaiLandrePreview from "./KaiLandrePreview";
-import ReactPlayer from 'react-player/lazy'
+import PreviewOrProjectPage from "./PreviewOrProjectPage";
 
 const Main = styled.main<FlexboxProps>`
   display: flex;
@@ -14,15 +16,25 @@ const Container = styled.div<LayoutProps>`
   ${layout};
 `;
 
-const KaiLandre = () => {
-  const url = 'https://www.youtube.com/watch?v=AGSH_acR4wA'
+const KaiLandreContent: React.FC = () => {
+  const url = "https://www.youtube.com/watch?v=AGSH_acR4wA";
 
   return (
+    <Container height="100%" width="100%">
+      <ReactPlayer url={url} width="100%" height="100%" />
+    </Container>
+  );
+};
+
+const launchDate = "2019-12-21";
+const KaiLandre: React.FC = () => {
+  return (
     <Main justifyContent="center" alignItems="center">
-      {/* <Container height='100%' width='100%'>
-      <ReactPlayer url={url} width='100%' height='100%'/>
-      </Container> */}
-      <KaiLandrePreview/>
+      <PreviewOrProjectPage
+        launchDate={launchDate}
+        PreviewPage={KaiLandrePreview}
+        ProjectPage={KaiLandreContent}
+      />
     </Main>
   );
 };
