+++ src/components/project-pages/KaiLandrePreview.tsx
@@ -4,8 +4,6 @@ import Timer from "../Timer";
 import { FlexboxProps, flexbox, layout, LayoutProps } from "styled-system";
 import dragon from "../assets/project-page/dragon.png";
 
-
-
 const Container = styled.div<FlexboxProps & LayoutProps>`
   ${flexbox};
   ${layout};
@@ -15,12 +13,20 @@ const Img = styled.img<LayoutProps>`
   ${layout};
 `;
 
-const KaiLandrePreview = () => {
+interface KaiLandrePreviewProps {
+  launchDate: string;
+}
+const KaiLandrePreview: React.FC<KaiLandrePreviewProps> = ({ launchDate }) => {
   return (
     <Fragment>
-      <Container flexDirection="column" display="flex" justifyContent="center" alignItems="center">
+      <Container
+        flexDirection="column"
+        display="flex"
+        justifyContent="center"
+        alignItems="center"
+      >
         <Img src={dragon} alt="dragon icon" maxWidth="30%" />
-        <Timer endDate="2020-12-21" />
+        <Timer launchDate={launchDate} />
       </Container>
     </Fragment>
   );
